[MOVE] Moved code to ol_account module
