+++ ls_pdf_reports/reports/__init__.py
@@ -1,8 +1,4 @@
 # from . import invoice_rma_refund
-from . import invoice_generic
-from . import invoice_proforma
-# TODO: sale.order.payment.method model not found
-# from . import invoice_proforma_from_so
 from . import packing_slip
 from . import packing_slip_from_so
 from . import picking_list
