+++ ol_account/__manifest__.py
@@ -18,5 +18,10 @@
     "data": [
         "security/security.xml",
         "views/res_config_settings_views.xml",
+        "views/account_invoice.xml",
+        "reports/invoice_generic.xml",
+        "reports/invoice_proforma_from_so.xml",
+        # TODO: sale.order.payment.method model not found
+        # "reports/invoice_proforma.xml",
     ],
 }
