+++ ol_account/reports/__init__.py
@@ -1,0 +1,3 @@
+from . import invoice_generic
+from . import invoice_proforma
+from . import invoice_proforma_from_so
