+++ ol_account/reports/invoice_proforma_from_so.xml
@@ -7,8 +7,8 @@
     <field name="print_report_name">'Proforma Invoice'</field>
     <field name="model">sale.order</field>
     <field name="report_type">qweb-pdf</field>
-    <field name="report_name">ls_pdf_reports.report_proforma_invoice_from_so</field>
-    <field name="report_file">ls_pdf_reports.report_proforma_invoice_from_so</field>
+    <field name="report_name">ol_account.report_proforma_invoice_from_so</field>
+    <field name="report_file">ol_account.report_proforma_invoice_from_so</field>
     <field name="binding_model_id" ref="sale.model_sale_order" />
     <field name="paperformat_xmlid_base">ls_templates.ls_report_paper_format</field>
   </record>
@@ -27,7 +27,7 @@
         <t t-set="sale_order" t-value="doc.with_context(lang=language)"/>
 
         <!-- Call the template document with the language set -->
-        <t t-call="ls_pdf_reports.report_sale_order_proforma_invoice_from_so_document" t-lang="language"/>
+        <t t-call="ol_account.report_sale_order_proforma_invoice_from_so_document" t-lang="language"/>
       </t>
     </t>
   </template>
@@ -36,7 +36,7 @@
   <template id="report_sale_order_proforma_invoice_from_so_document">
     <t t-foreach="payment_methods" t-as="payment_method">
       <t t-set="payment_method" t-value="payment_method.with_context(lang=language)"/>
-      <t t-call="ls_pdf_reports.report_sale_order_proforma_invoice_document" t-lang="language"/>
+      <t t-call="ol_account.report_sale_order_proforma_invoice_document" t-lang="language"/>
     </t>
   </template>
