+++ ol_account/views/account_invoice.xml
@@ -1,14 +1,17 @@
 <odoo>
 
     <record id="account_move_form" model="ir.ui.view">
-        <field name="name">ls_pdf_reports.account_move_form</field>
+        <field name="name">ol_account.account_move_form</field>
         <field name="model">account.move</field>
         <field name="inherit_id" ref="account.view_move_form" />
         <field name="arch" type="xml">
             <button name="action_invoice_sent" position="after">
                 <!-- Add Print button -->
-                <button name="action_invoice_print" class="oe_highlight"
-                    string="Print" type="object"
+                <button
+                    name="action_invoice_print"
+                    class="oe_highlight"
+                    string="Print"
+                    type="object"
                      />
                      <!-- invisible="state == 'draft' or move_type in ('out_invoice', 'out_refund')" -->
             </button>
